refactor path config properties
